aiops-platform/rca-engine/requirements.txt
fastapi
uvicorn
pydantic
kafka-python-ng==2.2.3
neo4j
openai
python-dotenv
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.1.0
mcp[cli]
httpx
langgraph-checkpoint-postgres
psycopg[binary]
psycopg-pool
